+++ sdk/keyvault/azure-security-keyvault-certificates/src/main/java/com/azure/security/keyvault/certificates/CertificateClientBuilder.java
@@ -196,16 +196,15 @@ public final class CertificateClientBuilder {
     }
 
     /**
-     * Sets the vault endpoint url to send HTTP requests to.
+     * Sets the vault endpoint URL to send HTTP requests to.
      *
      * @param vaultUrl The vault endpoint url is used as destination on Azure to send requests to. If you have a
-     * certificate identifier, use {@link KeyVaultCertificateIdentifier#parse(String)} to parse it and obtain the
+     * certificate identifier, create a new {@link KeyVaultCertificateIdentifier} to parse it and obtain the
      * {@code vaultUrl} and other information.
      *
      * @return The updated {@link CertificateClientBuilder} object.
      *
-     * @throws IllegalArgumentException If {@code vaultUrl} cannot be parsed into a valid URL.
-     * @throws NullPointerException If {@code vaultUrl} is {@code null}.
+     * @throws IllegalArgumentException if {@code vaultUrl} is null or it cannot be parsed into a valid URL.
      */
     public CertificateClientBuilder vaultUrl(String vaultUrl) {
         if (vaultUrl == null) {