+++ sdk/keyvault/azure-security-keyvault-keys/src/main/java/com/azure/security/keyvault/keys/cryptography/CryptographyAsyncClient.java
@@ -208,7 +208,7 @@ public class CryptographyAsyncClient {
      * portion of the key is used for encryption. This operation requires the keys/encrypt permission.
      *
      * <p>The {@link EncryptionAlgorithm encryption algorithm} indicates the type of algorithm to use for encrypting the
-     * specified {@code plainText}. Possible values for asymmetric keys include:
+     * specified {@code plaintext}. Possible values for asymmetric keys include:
      * {@link EncryptionAlgorithm#RSA1_5 RSA1_5}, {@link EncryptionAlgorithm#RSA_OAEP RSA_OAEP} and
      * {@link EncryptionAlgorithm#RSA_OAEP_256 RSA_OAEP_256}.
      *
@@ -226,16 +226,16 @@ public class CryptographyAsyncClient {
      * {@codesnippet com.azure.security.keyvault.keys.cryptography.CryptographyAsyncClient.encrypt#EncryptionAlgorithm-byte}
      *
      * @param algorithm The algorithm to be used for encryption.
-     * @param plainText The content to be encrypted.
+     * @param plaintext The content to be encrypted.
      * @return A {@link Mono} containing a {@link EncryptResult} whose {@link EncryptResult#getCipherText() cipher text}
      * contains the encrypted content.
      * @throws ResourceNotFoundException If the key cannot be found for encryption.
      * @throws UnsupportedOperationException If the encrypt operation is not supported or configured on the key.
-     * @throws NullPointerException If {@code algorithm} or {@code plainText} are {@code null}.
+     * @throws NullPointerException If {@code algorithm} or {@code plaintext} are {@code null}.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<EncryptResult> encrypt(EncryptionAlgorithm algorithm, byte[] plainText) {
-        return encrypt(new EncryptOptions(algorithm, plainText, null, null), null);
+    public Mono<EncryptResult> encrypt(EncryptionAlgorithm algorithm, byte[] plaintext) {
+        return encrypt(new EncryptParameters(algorithm, plaintext, null, null), null);
     }
 
     /**
@@ -245,7 +245,7 @@ public class CryptographyAsyncClient {
      * portion of the key is used for encryption. This operation requires the keys/encrypt permission.
      *
      * <p>The {@link EncryptionAlgorithm encryption algorithm} indicates the type of algorithm to use for encrypting the
-     * specified {@code plainText}. Possible values for asymmetric keys include:
+     * specified {@code plaintext}. Possible values for asymmetric keys include:
      * {@link EncryptionAlgorithm#RSA1_5 RSA1_5}, {@link EncryptionAlgorithm#RSA_OAEP RSA_OAEP} and
      * {@link EncryptionAlgorithm#RSA_OAEP_256 RSA_OAEP_256}.
      *
@@ -260,31 +260,31 @@ public class CryptographyAsyncClient {
      * <p><strong>Code Samples</strong></p>
      * <p>Encrypts the content. Subscribes to the call asynchronously and prints out the encrypted content details when
      * a response has been received.</p>
-     * {@codesnippet com.azure.security.keyvault.keys.cryptography.CryptographyAsyncClient.encrypt#EncryptOptions}
+     * {@codesnippet com.azure.security.keyvault.keys.cryptography.CryptographyAsyncClient.encrypt#EncryptParameters}
      *
-     * @param encryptOptions The parameters to use in the encryption operation.
+     * @param encryptParameters The parameters to use in the encryption operation.
      * @return A {@link Mono} containing a {@link EncryptResult} whose {@link EncryptResult#getCipherText() cipher text}
      * contains the encrypted content.
      * @throws ResourceNotFoundException If the key cannot be found for encryption.
      * @throws UnsupportedOperationException If the encrypt operation is not supported or configured on the key.
-     * @throws NullPointerException If {@code encryptOptions} is {@code null}.
+     * @throws NullPointerException If {@code encryptParameters} is {@code null}.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<EncryptResult> encrypt(EncryptOptions encryptOptions) {
-        Objects.requireNonNull(encryptOptions, "'encryptOptions' cannot be null");
+    public Mono<EncryptResult> encrypt(EncryptParameters encryptParameters) {
+        Objects.requireNonNull(encryptParameters, "'encryptParameters' cannot be null");
 
         try {
-            return withContext(context -> encrypt(encryptOptions, context));
+            return withContext(context -> encrypt(encryptParameters, context));
         } catch (RuntimeException ex) {
             return monoError(logger, ex);
         }
     }
 
 
-    Mono<EncryptResult> encrypt(EncryptOptions encryptOptions, Context context) {
+    Mono<EncryptResult> encrypt(EncryptParameters encryptParameters, Context context) {
         return ensureValidKeyAvailable().flatMap(available -> {
             if (!available) {
-                return cryptographyServiceClient.encrypt(encryptOptions, context);
+                return cryptographyServiceClient.encrypt(encryptParameters, context);
             }
 
             if (!checkKeyPermissions(this.key.getKeyOps(), KeyOperation.ENCRYPT)) {
@@ -292,7 +292,7 @@ public class CryptographyAsyncClient {
                     "Encrypt Operation is missing permission/not supported for key with id %s", key.getId()))));
             }
 
-            return localKeyCryptographyClient.encryptAsync(encryptOptions, context, key);
+            return localKeyCryptographyClient.encryptAsync(encryptParameters, context, key);
         });
     }
 
@@ -321,15 +321,15 @@ public class CryptographyAsyncClient {
      * {@codesnippet com.azure.security.keyvault.keys.cryptography.CryptographyAsyncClient.decrypt#EncryptionAlgorithm-byte}
      *
      * @param algorithm The algorithm to be used for decryption.
-     * @param cipherText The content to be decrypted.
+     * @param ciphertext The content to be decrypted.
      * @return A {@link Mono} containing the decrypted blob.
      * @throws ResourceNotFoundException If the key cannot be found for decryption.
      * @throws UnsupportedOperationException If the decrypt operation is not supported or configured on the key.
-     * @throws NullPointerException If {@code algorithm} or {@code cipherText} are {@code null}.
+     * @throws NullPointerException If {@code algorithm} or {@code ciphertext} are {@code null}.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<DecryptResult> decrypt(EncryptionAlgorithm algorithm, byte[] cipherText) {
-        return decrypt(new DecryptOptions(algorithm, cipherText, null, null, null));
+    public Mono<DecryptResult> decrypt(EncryptionAlgorithm algorithm, byte[] ciphertext) {
+        return decrypt(new DecryptParameters(algorithm, ciphertext, null, null, null));
     }
 
     /**
@@ -354,29 +354,29 @@ public class CryptographyAsyncClient {
      * <p><strong>Code Samples</strong></p>
      * <p>Decrypts the encrypted content. Subscribes to the call asynchronously and prints out the decrypted content
      * details when a response has been received.</p>
-     * {@codesnippet com.azure.security.keyvault.keys.cryptography.CryptographyAsyncClient.decrypt#DecryptOptions}
+     * {@codesnippet com.azure.security.keyvault.keys.cryptography.CryptographyAsyncClient.decrypt#DecryptParameters}
      *
-     * @param decryptOptions The parameters to use in the decryption operation.
+     * @param decryptParameters The parameters to use in the decryption operation.
      * @return A {@link Mono} containing the decrypted blob.
      * @throws ResourceNotFoundException If the key cannot be found for decryption.
      * @throws UnsupportedOperationException If the decrypt operation is not supported or configured on the key.
-     * @throws NullPointerException If {@code decryptOptions} is {@code null}.
+     * @throws NullPointerException If {@code decryptParameters} is {@code null}.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<DecryptResult> decrypt(DecryptOptions decryptOptions) {
-        Objects.requireNonNull(decryptOptions, "'decryptOptions' cannot be null");
+    public Mono<DecryptResult> decrypt(DecryptParameters decryptParameters) {
+        Objects.requireNonNull(decryptParameters, "'decryptParameters' cannot be null");
 
         try {
-            return withContext(context -> decrypt(decryptOptions, context));
+            return withContext(context -> decrypt(decryptParameters, context));
         } catch (RuntimeException ex) {
             return monoError(logger, ex);
         }
     }
 
-    Mono<DecryptResult> decrypt(DecryptOptions decryptOptions, Context context) {
+    Mono<DecryptResult> decrypt(DecryptParameters decryptParameters, Context context) {
         return ensureValidKeyAvailable().flatMap(available -> {
             if (!available) {
-                return cryptographyServiceClient.decrypt(decryptOptions, context);
+                return cryptographyServiceClient.decrypt(decryptParameters, context);
             }
 
             if (!checkKeyPermissions(this.key.getKeyOps(), KeyOperation.DECRYPT)) {
@@ -384,7 +384,7 @@ public class CryptographyAsyncClient {
                     "Decrypt Operation is not allowed for key with id %s", key.getId()))));
             }
 
-            return localKeyCryptographyClient.decryptAsync(decryptOptions, context, key);
+            return localKeyCryptographyClient.decryptAsync(decryptParameters, context, key);
         });
     }