+++ sdk/keyvault/azure-security-keyvault-keys/src/main/java/com/azure/security/keyvault/keys/cryptography/LocalCryptographyAsyncClient.java
@@ -48,7 +48,7 @@ public class LocalCryptographyAsyncClient {
      * portion of the key is used for encryption. This operation requires the keys/encrypt permission.
      *
      * <p>The {@link EncryptionAlgorithm encryption algorithm} indicates the type of algorithm to use for encrypting the
-     * specified {@code plainText}. Possible values for asymmetric keys include:
+     * specified {@code plaintext}. Possible values for asymmetric keys include:
      * {@link EncryptionAlgorithm#RSA1_5 RSA1_5}, {@link EncryptionAlgorithm#RSA_OAEP RSA_OAEP} and
      * {@link EncryptionAlgorithm#RSA_OAEP_256 RSA_OAEP_256}.
      *
@@ -64,14 +64,14 @@ public class LocalCryptographyAsyncClient {
      * {@codesnippet com.azure.security.keyvault.keys.cryptography.LocalCryptographyAsyncClient.encrypt#EncryptionAlgorithm-byte}
      *
      * @param algorithm The algorithm to be used for encryption.
-     * @param plainText The content to be encrypted.
+     * @param plaintext The content to be encrypted.
      * @return A {@link Mono} containing a {@link EncryptResult} whose {@link EncryptResult#getCipherText() cipher text}
      * contains the encrypted content.
      * @throws UnsupportedOperationException If the encrypt operation is not supported or configured on the key.
-     * @throws NullPointerException if {@code algorithm} or  {@code plainText} is {@code null}.
+     * @throws NullPointerException if {@code algorithm} or  {@code plaintext} is {@code null}.
      */
-    public Mono<EncryptResult> encrypt(EncryptionAlgorithm algorithm, byte[] plainText) {
-        return cryptographyAsyncClient.encrypt(algorithm, plainText);
+    public Mono<EncryptResult> encrypt(EncryptionAlgorithm algorithm, byte[] plaintext) {
+        return cryptographyAsyncClient.encrypt(algorithm, plaintext);
     }
 
     /**
@@ -81,7 +81,7 @@ public class LocalCryptographyAsyncClient {
      * portion of the key is used for encryption. This operation requires the keys/encrypt permission.
      *
      * <p>The {@link EncryptionAlgorithm encryption algorithm} indicates the type of algorithm to use for encrypting the
-     * specified {@code plainText}. Possible values for asymmetric keys include:
+     * specified {@code plaintext}. Possible values for asymmetric keys include:
      * {@link EncryptionAlgorithm#RSA1_5 RSA1_5}, {@link EncryptionAlgorithm#RSA_OAEP RSA_OAEP} and
      * {@link EncryptionAlgorithm#RSA_OAEP_256 RSA_OAEP_256}.
      *
@@ -94,16 +94,16 @@ public class LocalCryptographyAsyncClient {
      * <p><strong>Code Samples</strong></p>
      * <p>Encrypts the content. Subscribes to the call asynchronously and prints out the encrypted content details when
      * a response has been received.</p>
-     * {@codesnippet com.azure.security.keyvault.keys.cryptography.LocalCryptographyAsyncClient.encrypt#EncryptOptions}
+     * {@codesnippet com.azure.security.keyvault.keys.cryptography.LocalCryptographyAsyncClient.encrypt#EncryptParameters}
      *
-     * @param encryptOptions The parameters to use in the encryption operation.
+     * @param encryptParameters The parameters to use in the encryption operation.
      * @return A {@link Mono} containing a {@link EncryptResult} whose {@link EncryptResult#getCipherText() cipher text}
      * contains the encrypted content.
      * @throws UnsupportedOperationException If the encrypt operation is not supported or configured on the key.
-     * @throws NullPointerException if {@code encryptOptions} is {@code null}.
+     * @throws NullPointerException if {@code encryptParameters} is {@code null}.
      */
-    public Mono<EncryptResult> encrypt(EncryptOptions encryptOptions) {
-        return cryptographyAsyncClient.encrypt(encryptOptions);
+    public Mono<EncryptResult> encrypt(EncryptParameters encryptParameters) {
+        return cryptographyAsyncClient.encrypt(encryptParameters);
     }
 
     /**
@@ -129,13 +129,13 @@ public class LocalCryptographyAsyncClient {
      * {@codesnippet com.azure.security.keyvault.keys.cryptography.LocalCryptographyAsyncClient.decrypt#EncryptionAlgorithm-byte}
      *
      * @param algorithm The algorithm to be used for decryption.
-     * @param cipherText The content to be decrypted.
+     * @param ciphertext The content to be decrypted.
      * @return A {@link Mono} containing the decrypted blob.
      * @throws UnsupportedOperationException If the decrypt operation is not supported or configured on the key.
-     * @throws NullPointerException If {@code algorithm} or {@code cipherText} are {@code null}.
+     * @throws NullPointerException If {@code algorithm} or {@code ciphertext} are {@code null}.
      */
-    public Mono<DecryptResult> decrypt(EncryptionAlgorithm algorithm, byte[] cipherText) {
-        return cryptographyAsyncClient.decrypt(algorithm, cipherText);
+    public Mono<DecryptResult> decrypt(EncryptionAlgorithm algorithm, byte[] ciphertext) {
+        return cryptographyAsyncClient.decrypt(algorithm, ciphertext);
     }
 
     /**
@@ -158,15 +158,15 @@ public class LocalCryptographyAsyncClient {
      * <p><strong>Code Samples</strong></p>
      * <p>Decrypts the encrypted content. Subscribes to the call asynchronously and prints out the decrypted content
      * details when a response has been received.</p>
-     * {@codesnippet com.azure.security.keyvault.keys.cryptography.LocalCryptographyAsyncClient.decrypt#DecryptOptions}
+     * {@codesnippet com.azure.security.keyvault.keys.cryptography.LocalCryptographyAsyncClient.decrypt#DecryptParameters}
      *
-     * @param decryptOptions The parameters to use in the decryption operation.
+     * @param decryptParameters The parameters to use in the decryption operation.
      * @return A {@link Mono} containing the decrypted blob.
      * @throws UnsupportedOperationException If the decrypt operation is not supported or configured on the key.
-     * @throws NullPointerException If {@code decryptOptions} is {@code null}.
+     * @throws NullPointerException If {@code decryptParameters} is {@code null}.
      */
-    public Mono<DecryptResult> decrypt(DecryptOptions decryptOptions) {
-        return cryptographyAsyncClient.decrypt(decryptOptions);
+    public Mono<DecryptResult> decrypt(DecryptParameters decryptParameters) {
+        return cryptographyAsyncClient.decrypt(decryptParameters);
     }