+++ sdk/keyvault/azure-security-keyvault-secrets/src/main/java/com/azure/security/keyvault/secrets/SecretClientBuilder.java
@@ -191,10 +191,10 @@ public final class SecretClientBuilder {
     }
 
     /**
-     * Sets the vault url to send HTTP requests to.
+     * Sets the vault URL to send HTTP requests to.
      *
      * @param vaultUrl The vault url is used as destination on Azure to send requests to. If you have a secret
-     * identifier, use {@link KeyVaultSecretIdentifier#parse(String)} to parse it and obtain the {@code vaultUrl} and
+     * identifier, create a new {@link KeyVaultSecretIdentifier} to parse it and obtain the {@code vaultUrl} and
      * other information.
      *
      * @return The updated {@link SecretClientBuilder} object.