+++ sdk/keyvault/azure-security-keyvault-administration/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.0.0-beta.6 (Unreleased)
 
 ### New features
+- Added support for service version `7.2`.
 - Added support to specify whether or not a pipeline policy should be added per call or per retry.
 
 ## 4.0.0-beta.5 (2021-03-12)